% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_regression_estimates}
\alias{combine_regression_estimates}
\title{Combine regression estimates}
\usage{
combine_regression_estimates(
  est_list,
  est_name = "estimate",
  seq = 1:length(est_list)
)
}
\arguments{
\item{est_list}{A list of regression tibbles}

\item{est_name}{A character for the estimate column name, default is 'estimate'}

\item{seq}{a numeric vector to label each regression tibble, default is a sequential number;}
}
\value{
A tibble
}
\description{
This function combine a list of regression results
}
