% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{cox_summary}
\alias{cox_summary}
\title{Get cox coefficient summary table}
\usage{
cox_summary(fit, raw = F)
}
\arguments{
\item{raw}{if you want to output the raw summary table, default is False}

\item{df}{A cox fit}
}
\value{
A summary table in html format
}
\description{
This function gets coef summary table from a cox model, can be achieved by `get_regression_estimates`.
}
