% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{format_p}
\alias{format_p}
\title{Format p values}
\usage{
format_p(df, cut = 0.05, p_names = "p", col_names = NA)
}
\arguments{
\item{df}{A tibble or dataframe of the data}

\item{cut}{A number of the significance cut off, default is 0.05}

\item{p_names}{A character vector of the column names of p values, default is 'p'}

\item{col_names}{A character vector of output table column names, default is NA, which uses original names}
}
\value{
A html table
}
\description{
This function formats the p values in a table
}
