% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_contigency_table}
\alias{get_contigency_table}
\title{Get contigency table}
\usage{
get_contigency_table(
  df,
  left_var,
  right_var,
  test = F,
  highlight = F,
  highlight_p = 0.05
)
}
\arguments{
\item{df}{A tibble or dataframe of the data}

\item{left_var}{A character of the left variable name}

\item{right_var}{A character of the right variable name}

\item{test}{logical, T if you want to perfrom the test, default is F}

\item{highlight, }{logical, highlight=T highlights p values, default is F}

\item{highlight_p, }{a numeric value of the p value cutoff to highlight, default is 0.05}
}
\value{
A tibble with centered or standardized variables
}
\description{
This function genartes contigency table and performs mcnemar.test if requested
}
