% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{get_desc_stat_grouping}
\alias{get_desc_stat_grouping}
\title{Get Descriptive Stat for the whole sample and multiple subsets}
\usage{
get_desc_stat_grouping(
  dat,
  grouping,
  test = T,
  raw = F,
  median_vars = NULL,
  detail = F,
  detail_simple = F,
  highlight = F,
  highlight_p = 0.05,
  paired = F,
  sort = F
)
}
\arguments{
\item{dat}{A dataframe}

\item{grouping}{A string, the name of the grouping variable}

\item{test}{if you want to perform statistical testing, default is True}

\item{raw}{if you want to get dataframe raw output, default is False}

\item{median_vars}{specify the variables that need median(IQR) instead of mean(SD)}

\item{detail}{if you want to get all the descriptive stat for continuous variables, default is F}

\item{detail_simple}{if you  want to get all the descriptive stat in a condensed version for continuous variables, default is F}

\item{highlight}{logical, highlight=T highlights p values, default is F}

\item{highlight_p}{a numeric value of the p value cutoff to highlight, default is 0.05}

\item{paired}{logical, paired=T uses wilcoxon signed rank tests, default is F}

\item{sort}{logical, sort =T sorts the variables based on their sequence in the data, default is F, which puts continuous variables first}
}
\value{
A html table for descriptive statistics
}
\description{
This function generates descriptive stat for all the subsets in the sample based on the grouping variable.
The grouping variable cannot have NA. For all other cattegorical variables, NA will be replaced with Unknown.
}
