% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{get_group_compare}
\alias{get_group_compare}
\title{Test each variable in a df for multiple groups}
\usage{
get_group_compare(df, grouping, raw = F, format = T, paired = F)
}
\arguments{
\item{df}{A dataframe}

\item{grouping}{character, the column to separate the data into two subsets}

\item{raw}{logical, if you want to get table output, default is F}

\item{format}{logical, format = T outputs a formatted table, default is T}

\item{paired}{logical, paired=T uses wilcoxon signed rank tests, default is F}
}
\value{
A html table for simple univariate tests
}
\description{
This function generates wilcoxon rank rum test(for 2 groups) or  KW test(for more than 2 groups) for every continuous variable and
Chisq test(or Fisher test) for each categorical variable
}
