% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_regression_estimates}
\alias{get_regression_estimates}
\title{Get a summary of regression results}
\usage{
get_regression_estimates(
  df,
  outcome,
  predictor_vec,
  outcome_type,
  format = F,
  interaction = NULL,
  weights = NULL,
  tobit_upper = NULL,
  test_hypo = NULL,
  aic = F,
  highlight = F,
  highlight_p = 0.05
)
}
\arguments{
\item{df}{A dataframe}

\item{outcome}{A string, the name of the outcome}

\item{predictor_vec}{A character vector of predictors}

\item{outcome_type}{A string, can be one of these options:linear, binary, poisson, ordinal, tobit, normal_gee,
binary_gee, poisson_gee, ordinal_gee, lme}

\item{format}{logical, format=T outputs a fomatted table, only supported for logistic regression}

\item{interaction}{a character vector of interaction terms}

\item{weights}{a numeric vector of weights}

\item{tobit_upper}{a number, the upper bound of tobit models}

\item{test_hypo}{a linear hypothesis, only for linear regression}

\item{aic}{logical, aic=T also outputs AIC of the model, only for linear regression}

\item{highlight, }{logical, highlight=T highlights p values, default is F}

\item{highlight_p, }{a numeric value of the p value cutoff to highlight, default is 0.05}
}
\value{
A dataframe or a formatted html table
}
\description{
This function summarizes the results for multiple types of regression
}
