% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{get_row_number_indent}
\alias{get_row_number_indent}
\title{Get row numbers to indent}
\usage{
get_row_number_indent(dat, use_blank = F)
}
\arguments{
\item{dat}{a dataset with the first two columns, variable and stat}

\item{use_blank}{select rows based on a blank second column}
}
\value{
A vec of row numbers
}
\description{
This function returns the row numbers to indent
}
