% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sig}
\alias{get_sig}
\title{Get significant p values}
\usage{
get_sig(df, cut = 0.05, p_names = "p", any = T)
}
\arguments{
\item{df}{A tibble or a dataframe}

\item{cut}{A number of the cutoff to define significance, default is 0.05}

\item{p_names}{A character vector of column names of p values}

\item{any}{logical, if T, returns row with any significant p;
otherwise, return rows with all significant p s, default is T}
}
\value{
A tibble
}
\description{
This function returns rows with a significant p value defined by a cutoff
}
