% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_outlier}
\alias{is_outlier}
\title{Check outlier}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A logical vector
}
\description{
This function checks if each value in a vector is an outlier
}
