% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{na_to_unknown}
\alias{na_to_unknown}
\title{Recode NA}
\usage{
na_to_unknown(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe with NA recoded to Unknown
}
\description{
This function recodes NA to unknown for all the character columns.
Factor variables will be converted to character variables.
}
