% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{out_text}
\alias{out_text}
\title{Output variable names for R markdown}
\usage{
out_text(in_vec)
}
\arguments{
\item{in_vec}{A string vec}
}
\value{
A string
}
\description{
This function outputs variable list for markdown and save to the clipboard
}
