% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_characteristics.R
\name{plot_numeric_dist}
\alias{plot_numeric_dist}
\title{Plot distribution}
\usage{
plot_numeric_dist(df, grouping = NULL)
}
\arguments{
\item{df}{A dataframe}
}
\value{
histograms
}
\description{
This function plots the histogram of all the numeric variables in a df. A grouping variable can be provided.
}
