% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_results}
\alias{rename_results}
\title{rename regression results}
\usage{
rename_results(dat, n, est_name = "estimate")
}
\arguments{
\item{dat}{A tibble or a dataframe}

\item{n}{A character or a number as the suffix of renamed variables}

\item{est_name}{A character for the estimate column name, default is 'estimate'}
}
\value{
A tibble
}
\description{
This function renames regression results
}
