% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dat.R
\docType{data}
\name{sample_dat}
\alias{sample_dat}
\title{Sample Data}
\format{
a tibble(dataframe)
}
\usage{
data(sample_dat)
}
\description{
A sample dataset modified from the dataset starwars in tidyverse.
}
\details{
Generated by the following syntax:
sample_dat = starwars \%>\%  mutate(hair_color_group = ifelse(hair_color \%in\% c('black', 'brown', 'none'), hair_color, 'other')) \%>\% select(height, sex, gender, mass, hair_color_group) \%>\% filter(!is.na(gender))
}
\examples{
data(sample_dat)
head(sample_dat)
str(sample_dat)
}
\keyword{datasets}
