% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyreport-package.R
\docType{package}
\name{tidyreport-package}
\alias{tidyreport}
\alias{tidyreport-package}
\title{tidyreport: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lorenzoFabbri/tidyreport}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
