% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toupper_first}
\alias{toupper_first}
\title{Capitalize the first letter}
\usage{
toupper_first(string)
}
\arguments{
\item{string}{A string or a string vec}
}
\value{
A string or a string vec
}
\description{
This function capitalizes the first letter of a string or a string vec
}
