% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{get_regression_estimates}
\alias{get_regression_estimates}
\title{Get a summary of regression results}
\usage{
get_regression_estimates(
  df = NULL,
  outcome = NULL,
  predictor_vec = NULL,
  outcome_type,
  random_slope_var = NULL,
  lme_method = "REML",
  time_var = NULL,
  event_var = NULL,
  format = F,
  interaction = NULL,
  fit = NULL,
  weights = NULL,
  tobit_upper = NULL,
  test_hypo = NULL,
  aic = F,
  highlight = F,
  highlight_p = 0.05
)
}
\arguments{
\item{df}{A dataframe, optional if you provide fit}

\item{outcome}{A string, the name of the outcome, optional if you provide fit}

\item{predictor_vec}{A character vector of predictors, optional if you provide fit}

\item{outcome_type}{A string, can be one of these options:linear, binary, poisson, ordinal, tobit, normal_gee,
binary_gee, poisson_gee, ordinal_gee, lme(random intercept), lme2(random intercept and random slope), logistic_glme, cox}

\item{random_slope_var}{A string, the name of the random slope variable, optional}

\item{lme_method}{A string, can be 'REML' or "ML", default is "REML"}

\item{time_var}{A string, the name of the time variable for cox regression}

\item{event_var}{A string, the name of the event variable for cox regression}

\item{format}{logical, format=T outputs a formatted table, only supported for logistic regression}

\item{interaction}{a character vector of interaction terms, optional if you provide fit}

\item{fit}{A fitted model object, optional if you provide df, outcome and predictor_vec}

\item{weights}{a numeric vector of weights, optional if you provide fit}

\item{tobit_upper}{a number, the upper bound of tobit models, optional if you provide fit}

\item{test_hypo}{a linear hypothesis, only for linear regression}

\item{aic}{logical, aic=T also outputs AIC of the model, only for linear regression}

\item{highlight, }{logical, highlight=T highlights p values, default is F}

\item{highlight_p, }{a numeric value of the p value cutoff to highlight, default is 0.05}
}
\value{
A dataframe or a formatted html table
}
\description{
This function summarizes the results for multiple types of regression.
You can specify the model structure by providing the dataset, outcome and predictors.
Alternatively, you can directly provide the fitted model object.
}
