% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{out_text}
\alias{out_text}
\title{Output a string from a vector}
\usage{
out_text(in_vec, sep = "'")
}
\arguments{
\item{in_vec}{A vector}

\item{sep}{A character for the separating character, default is '(a single quotation mark)}
}
\value{
A string
}
\description{
This function outputs a string from a vector and save it to the clipboard
}
